// src/extensions/tsserver/protocol.d.ts
  interface FileRequestArgs {
    /**
     * The file for the request (absolute pathname required).
     */
    file: string;
    projectFileName?: string;
  }
// ...
  interface FileRangeRequestArgs extends FileRequestArgs {
    /**
     * The line number for the request (1-based).
     */
    startLine: number;
    /**
     * The character offset (on the line) for the request (1-based).
     */
    startOffset: number;
    /**
     * The line number for the request (1-based).
     */
    endLine: number;
    /**
     * The character offset (on the line) for the request (1-based).
     */
    endOffset: number;
  }
  /**
   * Instances of this interface specify errorcodes on a specific location in a sourcefile.
   */
  interface CodeFixRequestArgs extends FileRangeRequestArgs {
    /**
     * Errorcodes we want to get the fixes for.
     */
    errorCodes?: ReadonlyArray<number>;
  }
  interface GetCombinedCodeFixRequestArgs {
    scope: GetCombinedCodeFixScope;
    fixId: {};
  }
  interface GetCombinedCodeFixScope {
    type: "file";
    args: FileRequestArgs;
  }
// ...
  interface CodeEdit {
    /**
     * First character of the text span to edit.
     */
    start: Location;
    /**
     * One character past last character of the text span to edit.
     */
    end: Location;
    /**
     * Replace the span defined above with this string (may be
     * the empty string).
     */
    newText: string;
  }
  interface FileCodeEdits {
    fileName: string;
    textChanges: CodeEdit[];
  }
  interface CodeFixResponse extends Response {
    /** The code actions that are available */
    body?: CodeFixAction[];
  }
  interface CodeAction {
    /** Description of the code action to display in the UI of the editor */
    description: string;
    /** Text changes to apply to each file as part of the code action */
    changes: FileCodeEdits[];
    /** A command is an opaque object that should be passed to `ApplyCodeActionCommandRequestArgs` without modification.  */
    commands?: {}[];
  }
  interface CombinedCodeActions {
    changes: ReadonlyArray<FileCodeEdits>;
    commands?: ReadonlyArray<{}>;
  }
  interface CodeFixAction extends CodeAction {
    /** Short name to identify the fix, for use by telemetry. */
    fixName: string;
    /**
     * If present, one may call 'getCombinedCodeFix' with this fixId.
     * This may be omitted to indicate that the code fix can't be applied in a group.
     */
    fixId?: {};
    /** Should be present if and only if 'fixId' is. */
    fixAllDescription?: string;
  }